wip: move some things around and start setting up tests for Field
